回滚

+++ modules/zh_cn/6.py
@@ -10,13 +10,12 @@ from graia.ariadne.app import Ariadne
 from graia.ariadne.event.message import GroupMessage
 from graia.ariadne.message.element import Plain, At, Image
 from graia.ariadne.message.parser.base import MatchContent
-from graia.ariadne.model import Group
+from graia.ariadne.model import Group, MemberPerm
 from graia.ariadne.util.saya import listen, decorate
 from graia.saya import Channel
 
 import botfunc
 import cache_var
-import depen
 
 channel = Channel.current()
 channel.name("6榜")
@@ -214,8 +213,10 @@ async def six_top(app: Ariadne, group: Group, event: GroupMessage):
 
 @listen(GroupMessage)
 @decorate(MatchContent("6，闭嘴"))
-@decorate(depen.check_authority_op())
 async def no_six(app: Ariadne, group: Group, event: GroupMessage):
+    admins = await botfunc.get_all_admin()
+    if event.sender.id not in admins and event.sender.permission not in [MemberPerm.Administrator, MemberPerm.Owner]:
+        return
     if group.id not in cache_var.no_6:
         cache_var.no_6.append(group.id)
         await botfunc.run_sql("""INSERT INTO no_six VALUES (%s)""", (group.id,))
@@ -233,8 +234,10 @@ async def no_six(app: Ariadne, group: Group, event: GroupMessage):
 
 @listen(GroupMessage)
 @decorate(MatchContent("6，张嘴"))
-@decorate(depen.check_authority_op())
 async def yes_six(app: Ariadne, group: Group, event: GroupMessage):
+    admins = await botfunc.get_all_admin()
+    if event.sender.id not in admins:
+        return
     if group.id in cache_var.no_6:
         cache_var.no_6.remove(group.id)
         await botfunc.run_sql("""DELETE FROM no_six WHERE gid = %s""", (group.id,))
